[contrib/egrul] split archive reading into archives.py
Move BlobURL, the source-data constants, list_archives,
crawl_archive(s_for_date), and the intra-dump duplicate merging out of
generate.py into a new archives.py. Add list_archives_by_date() that
encapsulates the format-version routing so generate.py just gets back
a sorted [(date, [BlobURL])] list.
