+++ contrib/egrul/generate.py
@@ -1,14 +1,7 @@
-from collections import defaultdict
 from datetime import date, datetime, timedelta
-import os
 from pathlib import Path
-import tempfile
-from typing import Generator, Optional, List, Iterable, Dict
-from zipfile import ZipFile
-from dataclasses import dataclass
+from typing import Optional, Iterable
 
-import numpy as np
-import pandas
 from pyspark import Row, StorageLevel
 from pyspark.sql import SparkSession
 from pyspark.sql.dataframe import DataFrame
@@ -20,53 +13,17 @@ from pyspark.sql.functions import (
     explode_outer,
     explode,
 )
-from google.cloud.storage import Client  # type: ignore
 
-from egrul_xml import parse_xml
+from archives import (
+    LOCAL_BUCKET_CACHE_DIR,
+    crawl_archives_for_date,
+    list_archives_by_date,
+)
 from schema import company_record_schema
 from zavod import Context
 from zavod import Dataset
 
-LOCAL_BUCKET_CACHE_DIR = Path(
-    os.environ.get("LOCAL_BUCKET_CACHE_DIR", tempfile.gettempdir())
-)
-SOURCE_DATA_BUCKET_NAME = "egrul.opensanctions.org"
 PROCESSESED_PREFIX = "ru_egrul/processed/"
-
-# Format versions differ slightly but not enough to affect our parsing.
-# 406/407 overlap heavily; we switch to 407 at 2025-01-01.
-# 407/408 overlap Feb 10 – Mar 7 2026; we switch to 408 at 2026-03-01.
-SOURCE_DATA_PREFIX_406 = "egrul/EGRUL_406/"
-SOURCE_DATA_PREFIX_407 = "egrul/EGRUL_407/"
-SOURCE_DATA_PREFIX_408 = "egrul/EGRUL_408/"
-
-
-@dataclass
-class BlobURL:
-    """A wrapper around blob URLs that can be pickled for Spark workers."""
-
-    url: str
-
-    def _split_url(self) -> tuple[str, str]:
-        """Split the URL into bucket name and blob name."""
-        # URL format: gs://bucket-name/path/to/blob
-        assert self.url.startswith("gs://")
-        split = self.url[5:].split("/", 1)
-        assert len(split) == 2, f"Invalid Blob URL: {self.url}"
-        return split[0], split[1]
-
-    @property
-    def bucket_name(self) -> str:
-        """Extract the bucket name from the URL."""
-        return self._split_url()[0]
-
-    @property
-    def name(self) -> str:
-        """Extract the blob name from the URL."""
-        return self._split_url()[1]
-
-    def __str__(self) -> str:
-        return self.url
 
 
 def day_before(d: date) -> date:
@@ -166,92 +123,6 @@ def update_companies_from_new_companies(
     return merged.union(not_to_merge_old).union(not_to_merge_new)
 
 
-def merge_duplicate_company_records(df: DataFrame) -> DataFrame:
-    """Deduplicate companies that have the same ID in the given DataFrame."""
-
-    dupe_ids = df.groupBy("id").count().where(col("count") > 1).drop("count")
-    dupes = df.join(dupe_ids, on="id", how="inner")
-    non_dupes = df.join(dupe_ids, on="id", how="left_anti")
-
-    def _merge_companies(pdf: pandas.DataFrame) -> pandas.DataFrame:
-        # TODO(Leon Handreke): Implement a more sophisticated merge strategy. For now, all of them
-        # seem to be dissolved and reopened companies, some bureaucratic artifact.
-        # Magic from
-        # https://stackoverflow.com/questions/45469417/sort-by-column-sub-values-in-pandas
-        newest = pdf.iloc[
-            np.argsort([x.get("incorporation_date") for x in pdf.legal_entity])
-        ]
-        return newest[:1]
-
-    deduped = dupes.groupBy("id").applyInPandas(_merge_companies, company_record_schema)
-    # We do this fancy union stuff to avoid running expensive python on the non-dupes
-    return non_dupes.union(deduped)
-
-
-def get_local_archive_path(blob_url: BlobURL) -> Path:
-    return LOCAL_BUCKET_CACHE_DIR / str(blob_url.name)
-
-
-def crawl_archive(blob_url: BlobURL) -> Generator[dict, None, None]:
-    data_date = get_archive_date_from_blob_url(blob_url)
-    context = get_context(data_date)
-
-    local_archive_path = get_local_archive_path(blob_url)
-    # TODO: Since we cache persistently locally (for running on Leon's machine),
-    # maybe a checksum comparison with the remote blob would be a good idea.
-    if not os.path.exists(local_archive_path):
-        context.log.info("Downloading archive: %s" % blob_url)
-        client = Client()
-        bucket = client.get_bucket(blob_url.bucket_name)
-        blob = bucket.blob(blob_url.name)
-        # mkdir -p the directory for the archive
-        local_archive_path.parent.mkdir(parents=True, exist_ok=True)
-        blob.download_to_filename(local_archive_path)
-
-    context.log.info(
-        "Opening local archive: %s (cache of %s)" % (local_archive_path, blob_url)
-    )
-
-    try:
-        with ZipFile(local_archive_path, "r") as zip:
-            for name in zip.namelist():
-                if not name.lower().endswith(".xml"):
-                    continue
-                with zip.open(name, "r") as fh:
-                    for e in parse_xml(context, fh):
-                        yield e
-    finally:
-        # Don't clean up the temporary file, for now this is being run on Leon's machine and it's
-        # okay to just have them cached.
-        # os.unlink(local_archive_path)
-        pass
-
-
-def crawl_archives_for_date(
-    spark: SparkSession,
-    archive_date: date,
-    archives: List[BlobURL],
-) -> DataFrame:
-    table_name = archive_date.isoformat().replace("-", "_")
-    if spark.catalog.tableExists(table_name):
-        return spark.table(table_name)
-
-    # TODO: Parallelizing on XML files (inside the zips) instead of just the zips
-    # would speed up dataframe building for days that only have few archives a lot.
-    blob_rdd = spark.sparkContext.parallelize(archives)
-    parsed_rows_rdd = blob_rdd.flatMap(crawl_archive)
-    # Persist this expensive computation to avoid doing it multiple times during the following join
-    # https://spark.apache.org/docs/latest/rdd-programming-guide.html#which-storage-level-to-choose
-    df = spark.createDataFrame(parsed_rows_rdd, schema=company_record_schema).persist(
-        StorageLevel.DISK_ONLY
-    )
-    df = merge_duplicate_company_records(df)
-
-    df.write.saveAsTable(table_name, mode="overwrite")
-
-    return spark.table(table_name)
-
-
 def write_companies_df_to_csv(df: DataFrame, path_prefix: Path) -> None:
     """Write the companies DataFrame to CSV files to be emitted as FtM in the ru_egrul crawler.
 
@@ -332,41 +203,6 @@ def write_companies_df_to_csv(df: DataFrame, path_prefix: Path) -> None:
     all_legal_entities_df.write.csv(str(path_prefix / "legalentities"), **csv_options)
 
 
-def get_archive_date_from_blob_url(blob_url: BlobURL) -> date:
-    """Gets an archive date from the blob URL."""
-    # blob_url.name format: "egrul/EGRUL_406/01.01.2022_FULL/EGRUL_FULL_2022-01-01_214.zip"
-    path_parts = blob_url.name.split("/")
-    if len(path_parts) < 2:
-        raise ValueError(f"Invalid blob name format: {blob_url.name}")
-
-    dirname = path_parts[-2]  # Get the directory name before the zip file
-    # 01-01 has a _FULL suffix
-    dirname = dirname.rstrip("_FULL")
-    return datetime.strptime(dirname, "%d.%m.%Y").date()
-
-
-def aggregate_archives_by_date(
-    archive_blobs: Iterable[BlobURL],
-) -> Dict[date, List[BlobURL]]:
-    archives_by_date = defaultdict(list)
-    for archive_blob in archive_blobs:
-        archive_date = get_archive_date_from_blob_url(archive_blob)
-        archives_by_date[archive_date].append(archive_blob)
-    return archives_by_date
-
-
-def list_archives(bucket_name: str, prefix: str) -> List[BlobURL]:
-    """List all archive blobs from Google Cloud Storage and convert to BlobURL objects."""
-    client = Client()
-    bucket = client.get_bucket(bucket_name)
-
-    return [
-        BlobURL(f"gs://{bucket_name}/{blob.name}")
-        for blob in bucket.list_blobs(prefix=prefix)
-        if blob.name.endswith(".zip")
-    ]
-
-
 def merge_company_record_dfs(
     context: Context, records: Iterable[DataFrame]
 ) -> DataFrame:
@@ -397,32 +233,8 @@ def crawl(context: Context) -> None:
     spark.sparkContext.setCheckpointDir("env/spark-checkpoint")
     spark.sparkContext.setLogLevel("WARN")
 
-    archives_406 = [
-        a
-        for a in list_archives(SOURCE_DATA_BUCKET_NAME, SOURCE_DATA_PREFIX_406)
-        if get_archive_date_from_blob_url(a) < date(2025, 1, 1)
-    ]
-    archives_407 = [
-        a
-        for a in list_archives(SOURCE_DATA_BUCKET_NAME, SOURCE_DATA_PREFIX_407)
-        if date(2025, 1, 1) <= get_archive_date_from_blob_url(a) < date(2026, 3, 1)
-    ]
-    archives_408 = [
-        a
-        for a in list_archives(SOURCE_DATA_BUCKET_NAME, SOURCE_DATA_PREFIX_408)
-        if get_archive_date_from_blob_url(a) >= date(2026, 3, 1)
-    ]
-    archives = archives_406 + archives_407 + archives_408
-
-    archives_by_date = sorted(aggregate_archives_by_date(archives).items())
-    archives_by_date = [
-        (d, archives)
-        for d, archives in archives_by_date
-        # For debugging (or manual partial resume), process only part of the data
-        # if date(2022, 1, 1) <= d <= date(2022, 12, 31)
-        # Take 2022-01-01 as the starting point
-        if date(2022, 1, 1) <= d
-    ]
+    # For debugging (or manual partial resume), narrow the date range via start_date.
+    archives_by_date = list_archives_by_date()
 
     # Process the archives first to have them ready in the warehouse for the
     # iterative join later.
